Updates following the comments
Updated constraints, renamed the component, got rid of fixed text sizes, consolidated the init methods to avoid repetitive logic and more
